% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_salaries.R
\name{get_salaries}
\alias{get_salaries}
\title{Get Salaries}
\usage{
get_salaries(
  years = 2008:2021,
  union_rate = "tv-l",
  area = "west",
  resolution = "monthly",
  dbg = TRUE
)
}
\arguments{
\item{years}{years of tariffs to try to download (default: 2008:2021)}

\item{union_rate}{default: "tv-l"}

\item{area}{default: "west"}

\item{resolution}{"monthly": Monatswerte, "annual": Jahreswerte, "quarterly":
Quartalswerte, "weekly": Wochenwerte, "weekdays": werktagswerte, "daily":
Tageswerte. Default: "monthly"}

\item{dbg}{print debug messages (default: TRUE)}
}
\value{
tibble with salaries for multiple years
}
\description{
wrapper around \code{\link[=get_salary]{get_salary()}} for multiple years
}
\examples{
salaries <- get_salaries(years = 2008:2021)
salaries
}
