% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_salaries.R
\name{get_salary}
\alias{get_salary}
\title{Get Salary}
\usage{
get_salary(year = 2009, union_rate = "tv-l", area = "west", type = 1)
}
\arguments{
\item{year}{year of tariff (default: 2009)}

\item{union_rate}{default: "tv-l"}

\item{area}{default: "west"}

\item{type}{1: Monatswerte, 12: Jahreswerte, 3: Quartalswerte, "w": Wochenwerte,
"t": Tageswerte, "v": werktagswerte (default: 1)}
}
\value{
tibble in list format with salaries
}
\description{
Get Salary
}
\examples{
get_salary()
}
