% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yearly_salary_and_bonus.R
\name{get_yearly_salary_and_bonus}
\alias{get_yearly_salary_and_bonus}
\title{Get Yearly Salary and Bonus}
\usage{
get_yearly_salary_and_bonus(
  union_rate,
  area,
  year,
  pay_group,
  step,
  data = NULL
)
}
\arguments{
\item{union_rate}{default: "tv-l"}

\item{area}{default: "west"}

\item{year}{year of tariff (default: 2009)}
}
\description{
Get Yearly Salary and Bonus
}
