% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on.R
\name{on}
\alias{on}
\alias{off}
\title{Turn TV on and off}
\usage{
on(verbose = FALSE, port = NULL)

off(verbose = FALSE)
}
\arguments{
\item{verbose}{talk to the user}

\item{port}{port number, if \code{NULL}, use a random port.}
}
\description{
Turn TV on and off
}
\examples{
\dontrun{
tv::on()
iris

tv::off()
}
}
