% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R, R/tv.R
\name{tv_count}
\alias{tv_count}
\alias{tv_any}
\alias{tv_lvcf}
\alias{tv_ts}
\alias{tv_min}
\alias{tv_max}
\alias{tv_mean}
\alias{tv_sum}
\alias{tv}
\alias{time_varying}
\alias{check_tv_data}
\alias{check_tv_exposure}
\alias{check_tv_specs}
\title{Create a time-varying dataset}
\usage{
tv_count(x, ...)

tv_any(x, ...)

tv_lvcf(x, ...)

tv_ts(x, current_time, ...)

tv_min(x, ...)

tv_max(x, ...)

tv_mean(x, ...)

tv_sum(x, ...)

time_varying(
  x,
  specs,
  exposure,
  ...,
  time_units = c("days", "seconds"),
  id = "pat_id",
  n_cores = 1
)

check_tv_data(x, time_units, id)

check_tv_exposure(x, expected_ids, time_units, id)

check_tv_specs(specs, expected_features = NULL)
}
\arguments{
\item{x}{A data.frame with four columns: <id>, "feature", "datetime", "value"}

\item{...}{Other arguments (not used at this time)}

\item{current_time}{The current grid row's time}

\item{specs}{a data.frame with four columns: "feature", "use_for_grid", "lookback_start", "lookback_end", "aggregation". See details below.}

\item{exposure}{a data.frame with (at least) three columns: <id>, "exposure_start", "exposure_stop"}

\item{time_units}{What time units should be used? Seconds or days}

\item{id}{The id to use. Default is "pat_id"}

\item{n_cores}{Number of cores to use}

\item{expected_ids}{A vector of expected ids based on the data.}

\item{expected_features}{A vector of expected features based on the data.}
}
\description{
Create a time-varying dataset
}
\details{
The defaults for specs are to use everything for the grid creation, and to set \code{lookback_start=0}, with a message in both cases.
  Currently supported aggregation functions include counting ("count" or "n"), last-value-carried forward ("last value" or "lvcf"),
  any/none ("any" or "binary"), time since ("time since" or "ts"), min/max/mean, and the special "event" (for which lookbacks are ignored).
}
\examples{
  data(tv_example)
  time_varying(tv_example$data, tv_example$specs, tv_example$exposure,
               time_units = "days", id = "mcn")
}
