% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{tv_aggregation}
\alias{tv_aggregation}
\alias{tv_count}
\alias{tv_any}
\alias{tv_lvcf}
\alias{tv_ts}
\alias{tv_min}
\alias{tv_max}
\alias{tv_mean}
\alias{tv_median}
\alias{tv_sum}
\title{Time-varying aggregation functions}
\usage{
tv_count(value, ...)

tv_any(value, ...)

tv_lvcf(value, datetime, ...)

tv_ts(datetime, current_time, ...)

tv_min(value, ...)

tv_max(value, ...)

tv_mean(value, ...)

tv_median(value, ...)

tv_sum(value, ...)
}
\arguments{
\item{value}{A vector of values}

\item{...}{Other arguments (not used at this time)}

\item{datetime}{A datetime}

\item{current_time}{The current grid row's time}
}
\value{
A scalar, indicating the corresponding aggregation over \code{value} or \code{datetime}.
}
\description{
Time-varying aggregation functions
}
