% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_example.R
\docType{data}
\name{tv_example}
\alias{tv_example}
\title{Example data for time-varying}
\format{
A list
\describe{
  \item{data}{The data}
  \item{specs}{The specs}
}
}
\usage{
tv_example
}
\description{
Example data for time-varying
}
\seealso{
\code{\link{tv}}
}
\keyword{datasets}
