% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_salaries.R
\name{get_salaries}
\alias{get_salaries}
\title{Get Salaries}
\usage{
get_salaries(
  years = 2000:2022,
  union_rate = "tv-l",
  area = "west",
  type = 1,
  dbg = TRUE
)
}
\arguments{
\item{years}{years of tariffs to try to download (default: 2000:2022)}

\item{union_rate}{default: "tv-l"}

\item{area}{default: "west"}

\item{type}{1: Monatswerte, 12: Jahreswerte, 3: Quartalswerte, "w": Wochenwerte,
"t": Tageswerte, "v": werktagswerte (default: 1)}

\item{dbg}{print debug}
}
\value{
tibble with salaries for multiple years
}
\description{
wrapper around \code{\link[=get_salary]{get_salary()}} for multiple years
}
\examples{
salaries <- get_salaries(years = 2010:2020)
salaries
}
