% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_salaries.R
\name{get_salary}
\alias{get_salary}
\title{Get Salary}
\usage{
get_salary(
  year = 2009L,
  union_rate = "tv-l",
  area = "west",
  resolution = "monthly"
)
}
\arguments{
\item{year}{year of tariff (default: 2009)}

\item{union_rate}{default: "tv-l"}

\item{area}{default: "west"}

\item{resolution}{"monthly": Monatswerte, "annual": Jahreswerte, "quarterly":
Quartalswerte, "weekly": Wochenwerte, "weekdays": werktagswerte, "daily":
Tageswerte. Default: "monthly"}
}
\value{
tibble in list format with salaries
}
\description{
Get Salary
}
\examples{
get_salary()
}
