% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{tv_aggregation}
\alias{tv_aggregation}
\alias{tv_count}
\alias{tv_any}
\alias{tv_lvcf}
\alias{tv_ts}
\alias{tv_min}
\alias{tv_max}
\alias{tv_mean}
\alias{tv_sum}
\title{Time-varying aggregation functions}
\usage{
tv_count(x, ...)

tv_any(x, ...)

tv_lvcf(x, ...)

tv_ts(x, current_time, ...)

tv_min(x, ...)

tv_max(x, ...)

tv_mean(x, ...)

tv_sum(x, ...)
}
\arguments{
\item{x}{A data.frame with four columns: <id>, "feature", "datetime", "value"}

\item{...}{Other arguments (not used at this time)}

\item{current_time}{The current grid row's time}
}
\description{
Time-varying aggregation functions
}
